core: rename parsing to parser for a better backward compat
