+++ enaml/core/parser/lexer2.py
