+++ tests/test_examples.py
@@ -14,7 +14,7 @@ import pytest
 from future.utils import exec_
 
 from enaml import imports
-from enaml.core.parsing import parse
+from enaml.core.parser import parse
 from enaml.core.enaml_compiler import EnamlCompiler
 from enaml.widgets.api import PopupView, Window
 from utils import (close_window_or_popup, get_popup,
